init distfree mode
