+++ src/main/python/plotlyst/view/style/base.py
@@ -24,7 +24,7 @@ from PyQt6.QtGui import QColor
 from PyQt6.QtWidgets import QWidget
 from qtmenu import MenuWidget
 
-from plotlyst.view.style.theme import BG_PRIMARY_COLOR, BG_SECONDARY_COLOR, BG_MUTED_COLOR, ALT_BG_COLOR
+from plotlyst.view.style.theme import BG_PRIMARY_COLOR, BG_SECONDARY_COLOR, BG_MUTED_COLOR, BG_ALT_COLOR
 
 style = f'''
 * {{
@@ -102,7 +102,7 @@ QFrame[muted-bg=true] {{
 }}
 
 QFrame[alt-bg=true] {{
-    background-color: {ALT_BG_COLOR};
+    background-color: {BG_ALT_COLOR};
 }}
 
 QFrame[highlighted-bg=true] {{
